+++ packages/frontend/src/pages/Dashboard.tsx
@@ -4,6 +4,7 @@ import {
     type DashboardTile,
     type Dashboard as IDashboard,
 } from '@lightdash/common';
+import { MantineProvider } from '@mantine-8/core';
 import { Box, Button, Flex, Group, Modal, Stack, Text } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 import { captureException, useProfiler } from '@sentry/react';
@@ -13,6 +14,7 @@ import { type Layout } from 'react-grid-layout';
 import { useBlocker, useNavigate, useParams } from 'react-router';
 import DashboardFilter from '../components/DashboardFilter';
 import DashboardTabs from '../components/DashboardTabs';
+import PinnedParameters from '../components/PinnedParameters';
 import DashboardHeader from '../components/common/Dashboard/DashboardHeader';
 import ErrorState from '../components/common/ErrorState';
 import MantineIcon from '../components/common/MantineIcon';
@@ -23,7 +25,7 @@ import DashboardDuplicateModal from '../components/common/modal/DashboardDuplica
 import { DashboardExportModal } from '../components/common/modal/DashboardExportModal';
 import { useDashboardCommentsCheck } from '../features/comments';
 import { DateZoom } from '../features/dateZoom';
-import { Parameters, useParameters } from '../features/parameters';
+import { Parameters } from '../features/parameters';
 import {
     appendNewTilesToBottom,
     useUpdateDashboard,
@@ -33,6 +35,7 @@ import { useOrganization } from '../hooks/organization/useOrganization';
 import useToaster from '../hooks/toaster/useToaster';
 import { useContentAction } from '../hooks/useContent';
 import { useSpaceSummaries } from '../hooks/useSpaces';
+import { getMantine8ThemeOverride } from '../mantine8Theme';
 import useApp from '../providers/App/useApp';
 import DashboardProvider from '../providers/Dashboard/DashboardProvider';
 import useDashboardContext from '../providers/Dashboard/useDashboardContext';
@@ -89,9 +92,6 @@ const Dashboard: FC = () => {
         (c) => c.setDashboardTemporaryFilters,
     );
     const isDateZoomDisabled = useDashboardContext((c) => c.isDateZoomDisabled);
-    const dashboardParameterReferences = useDashboardContext(
-        (c) => c.dashboardParameterReferences,
-    );
     const areAllChartsLoaded = useDashboardContext((c) => c.areAllChartsLoaded);
     const missingRequiredParameters = useDashboardContext(
         (c) => c.missingRequiredParameters,
@@ -105,10 +105,6 @@ const Dashboard: FC = () => {
     );
     const parameterValues = useDashboardContext((c) => c.parameterValues);
     const clearAllParameters = useDashboardContext((c) => c.clearAllParameters);
-    const setParameterDefinitions = useDashboardContext(
-        (c) => c.setParameterDefinitions,
-    );
-
     const hasDateZoomDisabledChanged = useMemo(() => {
         return (
             (dashboard?.config?.isDateZoomDisabled || false) !==
@@ -119,6 +115,33 @@ const Dashboard: FC = () => {
     const dashboardParameters = useDashboardContext(
         (c) => c.dashboardParameters,
     );
+    const pinnedParameters = useDashboardContext((c) => c.pinnedParameters);
+    const toggleParameterPin = useDashboardContext((c) => c.toggleParameterPin);
+    const havePinnedParametersChanged = useDashboardContext(
+        (c) => c.havePinnedParametersChanged,
+    );
+    const setHavePinnedParametersChanged = useDashboardContext(
+        (c) => c.setHavePinnedParametersChanged,
+    );
+    const setPinnedParameters = useDashboardContext(
+        (c) => c.setPinnedParameters,
+    );
+
+    const parameterDefinitions = useDashboardContext(
+        (c) => c.parameterDefinitions,
+    );
+
+    const parameterReferences = useDashboardContext(
+        (c) => c.dashboardParameterReferences,
+    );
+
+    const referencedParameters = useMemo(() => {
+        return Object.fromEntries(
+            Object.entries(parameterDefinitions).filter(([key]) =>
+                parameterReferences.has(key),
+            ),
+        );
+    }, [parameterDefinitions, parameterReferences]);
 
     const {
         enabled: isFullScreenFeatureEnabled,
@@ -157,20 +180,6 @@ const Dashboard: FC = () => {
     const tabsEnabled = dashboardTabs && dashboardTabs.length > 0;
 
     const defaultTab = dashboardTabs?.[0];
-
-    const { data: parameterDefinitions } = useParameters(
-        projectUuid,
-        Array.from(dashboardParameterReferences ?? []),
-        {
-            enabled: !!projectUuid && !!dashboardParameterReferences,
-        },
-    );
-
-    useEffect(() => {
-        if (parameterDefinitions) {
-            setParameterDefinitions(parameterDefinitions);
-        }
-    }, [parameterDefinitions, setParameterDefinitions]);
 
     useEffect(() => {
         if (isDashboardLoading) return;
@@ -275,6 +284,7 @@ const Dashboard: FC = () => {
         if (isSuccess) {
             setHaveTilesChanged(false);
             setHaveFiltersChanged(false);
+            setHavePinnedParametersChanged(false);
             setDashboardTemporaryFilters({
                 dimensions: [],
                 metrics: [],
@@ -302,6 +312,7 @@ const Dashboard: FC = () => {
         setDashboardTemporaryFilters,
         setHaveFiltersChanged,
         setHaveTilesChanged,
+        setHavePinnedParametersChanged,
         dashboardTabs,
         activeTab,
     ]);
@@ -449,6 +460,8 @@ const Dashboard: FC = () => {
         setHaveTabsChanged(false);
         setDashboardTabs(dashboard.tabs);
         setSavedParameters(dashboard.parameters ?? {});
+        setPinnedParameters(dashboard.config?.pinnedParameters ?? []);
+        setHavePinnedParametersChanged(false);
 
         if (dashboardTabs.length > 0) {
             void navigate(
@@ -475,6 +488,8 @@ const Dashboard: FC = () => {
         dashboardTabs,
         activeTab,
         setSavedParameters,
+        setPinnedParameters,
+        setHavePinnedParametersChanged,
     ]);
 
     const handleMoveDashboardToSpace = useCallback(
@@ -564,7 +579,7 @@ const Dashboard: FC = () => {
     }
 
     return (
-        <>
+        <MantineProvider theme={getMantine8ThemeOverride()}>
             {blocker.state === 'blocked' && (
                 <Modal
                     opened
@@ -631,7 +646,8 @@ const Dashboard: FC = () => {
                             hasTemporaryFilters ||
                             haveTabsChanged ||
                             hasDateZoomDisabledChanged ||
-                            parametersHaveChanged
+                            parametersHaveChanged ||
+                            havePinnedParametersChanged
                         }
                         onAddTiles={handleAddTiles}
                         onSaveDashboard={() => {
@@ -669,6 +685,7 @@ const Dashboard: FC = () => {
                                 tabs: dashboardTabs,
                                 config: {
                                     isDateZoomDisabled,
+                                    pinnedParameters,
                                 },
                                 parameters: dashboardParameters,
                             });
@@ -706,23 +723,28 @@ const Dashboard: FC = () => {
                     {/* DateZoom section will adjust width dynamically */}
                     {hasDashboardTiles && (
                         <Group spacing="xs" style={{ marginLeft: 'auto' }}>
-                            <Parameters
-                                isEditMode={isEditMode}
-                                parameterValues={parameterValues}
-                                onParameterChange={handleParameterChange}
-                                onClearAll={clearAllParameters}
-                                parameterReferences={
-                                    dashboardParameterReferences
-                                }
-                                areAllChartsLoaded={areAllChartsLoaded}
-                                missingRequiredParameters={
-                                    missingRequiredParameters
-                                }
-                            />
                             <DateZoom isEditMode={isEditMode} />
                         </Group>
                     )}
                 </Group>
+                {hasDashboardTiles && (
+                    <Group spacing="xs" align="flex-start" noWrap px={'lg'}>
+                        <Parameters
+                            isEditMode={isEditMode}
+                            parameterValues={parameterValues}
+                            onParameterChange={handleParameterChange}
+                            onClearAll={clearAllParameters}
+                            parameters={referencedParameters}
+                            isLoading={!areAllChartsLoaded}
+                            missingRequiredParameters={
+                                missingRequiredParameters
+                            }
+                            pinnedParameters={pinnedParameters}
+                            onParameterPin={toggleParameterPin}
+                        />
+                        <PinnedParameters isEditMode={isEditMode} />
+                    </Group>
+                )}
                 <Flex style={{ flexGrow: 1, flexDirection: 'column' }}>
                     <DashboardTabs
                         isEditMode={isEditMode}
@@ -774,7 +796,7 @@ const Dashboard: FC = () => {
                     />
                 )}
             </Page>
-        </>
+        </MantineProvider>
     );
 };