+++ packages/frontend/src/providers/Dashboard/types.ts
@@ -107,6 +107,11 @@ export type DashboardContextType = {
     tileParameterReferences: Record<string, string[]>;
     areAllChartsLoaded: boolean;
     parameterDefinitions: ParameterDefinitions;
-    setParameterDefinitions: Dispatch<SetStateAction<ParameterDefinitions>>;
+    addParameterDefinitions: (parameters: ParameterDefinitions) => void;
     missingRequiredParameters: string[];
+    pinnedParameters: string[];
+    setPinnedParameters: (parameters: string[]) => void;
+    toggleParameterPin: (parameterKey: string) => void;
+    havePinnedParametersChanged: boolean;
+    setHavePinnedParametersChanged: Dispatch<SetStateAction<boolean>>;
 };